Add GPT-2 model with start token
The start token allows to generate sequences just by giving that token
as input. In the conf the block size only refers to codes, whereas no.
positions is affected by both the codes and the start token. The
vocabulary size does not consider the start token either, it is
automatically added in the model. When generating, as in the training
the start token is never predicted its probability is very low so it is
never generated.